refactor(backend): split learning store bootstrap seed modules
